requirements.txt
anthropic>=0.50.0
langgraph>=0.2.0
langchain-core>=0.3.0
python-telegram-bot>=21.0
aiosqlite>=0.20.0
apscheduler>=3.10.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
pyyaml>=6.0.0
urllib3<2
# Big data & analytics
pandas>=2.0.0
duckdb>=0.10.0
polars>=0.20.0
openpyxl>=3.1.0
matplotlib>=3.8.0
seaborn>=0.13.0
xlsxwriter>=3.1.0
pyarrow>=15.0.0
# System monitoring
psutil>=5.9.0
# Testing
pytest-asyncio>=0.23.0
# Internet access & scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
duckduckgo-search>=6.0.0
httpx>=0.27.0
trafilatura>=1.8.0
